Add CHANGELOG generation during module build.
Update the .gemfile to source gems from https://rubygems.org to fix a
Travis/bundler warning.
Update the .gemfile to install puppet-lint.
Update the Modulefile summary and description.
Minor changes to the README.md to clarify what the module is and does
and sync with the Modulefile description.
Update the .fixtures.yml to test against puppetlabs-stdlib 2.6.0.
Update .travis.yml to support updated Puppet versions.
Update the Rakefile to add support for task :validate and task :lint.